# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GrowthAgent API",
    "version": "1.0.0",
    "description": "Autonomous AI growth system with ONE deterministic gatekeeper. Built for Razorpay AI Buildathon."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development & demo server"
    }
  ],
  "paths": {
...
    "/v1/stream-tickets": {
      "post": {
        "summary": "Mint SSE Stream Ticket",
        "description": "Exchange X-Agent-Key for a short-lived SSE ticket for browser EventSource.",
        "security": [
          {
            "AgentKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket minted with expiration."
          },
          "429": {
            "description": "Rate limited (RATE_LIMITED_HTTP)."
          }
        }
      }
    },
    "/v1/stream/{txId}": {
      "get": {
        "summary": "Server-Sent Events Transaction Audit Stream",
        "description": "SSE stream projecting the hash-chained audit log for one transaction. Admission is checked BEFORE the first byte, so a refusal is an ordinary JSON error rather than a half-open stream.",
        "responses": {
          "200": {
            "description": "SSE event stream."
          },
          "401": {
            "description": "Missing/invalid ticket or agent key."
          },
          "404": {
            "description": "Transaction not found or foreign agent."
          },
          "429": {
            "description": "Concurrent-stream ceiling reached (RATE_LIMITED_HTTP); details carry the global and per-agent limits."
          }
        }
      }
    },
    "/v1/webhooks/razorpay": {
      "post": {
        "summary": "Razorpay Payment Webhook Ingress",
        "description": "Raw HMAC-verified ingress moving transactions from AWAITING_PAYMENT to PAID.",
        "security": [
          {
            "RazorpaySignature": []
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook verified and processed."
          },
          "401": {
            "description": "Invalid HMAC signature."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-Key"
      },
      "AdminTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      },
      "RazorpaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Razorpay-Signature"
      }
    }
  }
}